Update attack_layer_builder.py
This change gives users the option to either manually input the technique IDs or select a CSV file containing the IDs. It also properly utilizes a single Tkinter root instance for both opening and saving files. Additionally, correctly placed the root.destroy() call to properly close the Tkinter window after use, for resource management. Finally, included a prompt for the user to input the version number of the ATT&CK Navigator they are using. A while loop is used to continuously prompt the user until a valid input (a number between 4 and 14) is entered.