Update attack_layer_builder.py
Introduces a default version setting for the ATT&CK Navigator. Users can now enter a version number, press Enter to accept the default version (v14), or be reverted to the default if the input is invalid or out of the expected range (4-14). The input is then formatted appropriately for JSON compatibility.